выполнение финального задания 6 спринта.(3.0)
