+++ src/com/model/tasks/Task.java
@@ -1,6 +1,6 @@
 package com.model.tasks;
 
-import com.status.status.Status;
+import com.status.Status;
 
 import java.util.Objects;
